agents/requirements.txt
# UNREDACTED AI Agent Service — Python dependencies
# Pinned for Railway deployment stability

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# LangChain + LangGraph AI stack
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-core==0.3.29
langgraph==0.2.62

# Anthropic SDK (used directly by langchain-anthropic)
anthropic==0.45.0

# HTTP client (async, used in all agents)
httpx==0.28.1

# Data validation
pydantic==2.10.6

# Environment config
python-dotenv==1.0.1
